Make memory sizes in struct device long
Some memory sizes are 2^16 or greater, and storing that value in an int
is not portable. Make relevant struct fields longs (at least 32 bits)
instead.

I sense this non-portable use of int issue is present in many other
places as well...

+++ src/avra.h
@@ -119,10 +119,10 @@ extern const int SEG_BSS_DATA;
 struct segment_info {
 	const char *name;
 	char ident;	  /* C, D, E */
-	int addr;     /* address in cells */
-	int count;    /* length in cells  */
-	int lo_addr;  /* lowest addr supported */
-	int hi_addr;  /* one past highest addr supported */
+	long addr;    /* address in cells */
+	long count;   /* length in cells  */
+	long lo_addr; /* lowest addr supported */
+	long hi_addr; /* one past highest addr supported */
 	int	cellsize; /* 1 for data/eeprom, 2 for flash  */
 	int flags;	  /* SEG_BSS_DATA */
